removed old ModelDataObject that was working on xarray datasets and replaced it eventually by the new version working od dictionaries: this one works well with dask and zarr
